splitting of the MIT stuff
