+++ crossbar/crossbar/adapter/rest/test/__init__.py
@@ -28,11 +28,20 @@
 #
 #####################################################################################
 
-from random import randint
+from __future__ import absolute_import
+
+import hmac
+import hashlib
+import random
+import base64
+
+from datetime import datetime
 
 from collections import namedtuple
 
 from twisted.internet.defer import maybeDeferred
+
+from crossbar.adapter.rest.test.requestMock import _requestMock, _render
 
 publishedMessage = namedtuple("pub", ["id"])
 
@@ -45,7 +54,7 @@ class MockPusherSession(object):
         self._published_messages = []
 
         def publish(topic, *args, **kwargs):
-            messageID = randint(0, 100000)
+            messageID = random.randint(0, 100000)
 
             self._published_messages.append({
                 "id": messageID,
@@ -60,3 +69,50 @@ class MockPusherSession(object):
             return maybeDeferred(publish, topic, *args, **kwargs)
 
         setattr(self, "publish", _publish)
+
+
+
+def _utcnow():
+    now = datetime.utcnow()
+    return now.strftime("%Y-%m-%dT%H:%M:%S.%f")[:-3] + "Z"
+
+
+def makeSignedArguments(params, signKey, signSecret, body):
+
+    params['timestamp'] = [_utcnow()]
+    params['seq'] = ["1"]
+    params['key'] = [signKey]
+    params['nonce'] = [str(random.randint(0, 9007199254740992))]
+
+    # HMAC[SHA256]_{secret} (key | timestamp | seq | nonce | body) => signature
+
+    hm = hmac.new(signSecret.encode('utf8'), None, hashlib.sha256)
+    hm.update(params['key'][0].encode('utf8'))
+    hm.update(params['timestamp'][0].encode('utf8'))
+    hm.update(u"{0}".format(params['seq'][0]).encode('utf8'))
+    hm.update(u"{0}".format(params['nonce'][0]).encode('utf8'))
+    hm.update(body)
+    signature = base64.urlsafe_b64encode(hm.digest())
+    params['signature'] = [signature]
+
+    return params
+
+
+def testResource(resource, path, params=None, method="GET", body="",
+                 headers=None, sign=False, signKey=None, signSecret=None):
+
+    params = {} if params == None else params
+    headers = {} if params == None else headers
+
+    def _cb(result, request):
+        return request
+
+    if sign:
+        params = makeSignedArguments(params, signKey, signSecret, body)
+
+    req = _requestMock(path, args=params, method=method,
+                       headers=headers, body=body)
+
+    d = _render(resource, req)
+    d.addCallback(_cb, req)
+    return d
