+++ crossbar/crossbar/adapter/rest/test/test_caller.py
@@ -36,7 +36,7 @@ from twisted.trial.unittest import TestCase
 from twisted.internet.defer import inlineCallbacks, maybeDeferred
 
 from crossbar.adapter.rest import CallerResource
-from crossbar.adapter.rest.test.requestMock import testResource
+from crossbar.adapter.rest.test import testResource
 
 
 class MockSession(object):
