+++ crossbar/crossbar/adapter/rest/test/test_signature.py
@@ -37,8 +37,7 @@ from twisted.internet.defer import inlineCallbacks, maybeDeferred
 
 from crossbar.adapter.rest.common import _CommonResource
 from crossbar.adapter.rest import PusherResource
-from crossbar.adapter.rest.test import MockPusherSession
-from crossbar.adapter.rest.test.requestMock import testResource
+from crossbar.adapter.rest.test import MockPusherSession, testResource
 
 
 class SignatureTestCase(TestCase):
